* Fixed the books and borrowed_books migrations. * Added fine watcher that will calculate fines and will be ran via task scheduler.

+++ database/migrations/2016_10_31_070444_create_borrowed_books_table.php
@@ -17,7 +17,7 @@ class CreateBorrowedBooksTable extends Migration
             $table->integer('user_id')->unsigned()->index();
             $table->integer('book_id')->unsigned()->index();
             $table->smallInteger('status')->unsigned()->default(2); //1 - borrowed, 0 - returned, 2 - pending/needs_approval
-            $table->decimal('fine', 5, 3)->default(0.0);
+            $table->double('fine', 15, 3)->default(0.0);
             $table->dateTime('borrow_start_date');
             $table->dateTime('date_returned');
             $table->timestamps();
